The 0 distance acts now like a reset.
TODO : the filters can only filter on the already filteredItems, which means that we can't filter less.

Solution idea : implement a private val List<Item> that is initialised by getAvailableLoans(), and on which the filters are applied (to avoid calling the db at each filter modification), and to update the UI on a filtered list.